fixup  kafka client
